Move test into its own file
